Remove old item templates & config apps, move starting equipment
Removes all of the unused item templates that were only used by
the old item sheet plus several partials that are no longer used.

Also moves `StartingEquipmentConfig` into a `config` folder to
match the folder structure for actors and removed the two stub
classes for `EnchantmentConfig` & `SummoningConfig`.
