+++ module/applications/item/_module.mjs
@@ -6,6 +6,4 @@ export {default as ItemSheet5e} from "./item-sheet.mjs";
 export {default as ItemSheet5e2} from "./item-sheet-2.mjs";
 export {default as ItemSheetV2Mixin} from "./sheet-v2-mixin.mjs";
 
-export {default as EnchantmentConfig} from "./enchantment-config.mjs";
-export {default as StartingEquipmentConfig} from "./starting-equipment-config.mjs";
-export {default as SummoningConfig} from "./summoning-config.mjs";
+export {default as StartingEquipmentConfig} from "./config/starting-equipment-config.mjs";
