+++ module/utils.mjs
@@ -780,17 +780,10 @@ export async function preloadHandlebarsTemplates() {
     "systems/dnd5e/templates/items/details/details-subclass.hbs",
     "systems/dnd5e/templates/items/details/details-tool.hbs",
     "systems/dnd5e/templates/items/details/details-weapon.hbs",
-    "systems/dnd5e/templates/items/parts/item-action.hbs",
-    "systems/dnd5e/templates/items/parts/item-activation.hbs",
     "systems/dnd5e/templates/items/parts/item-activities.hbs",
-    "systems/dnd5e/templates/items/parts/item-advancement.hbs",
     "systems/dnd5e/templates/items/parts/item-advancement2.hbs",
-    "systems/dnd5e/templates/items/parts/item-description.hbs",
     "systems/dnd5e/templates/items/parts/item-description2.hbs",
     "systems/dnd5e/templates/items/parts/item-details.hbs",
-    "systems/dnd5e/templates/items/parts/item-mountable.hbs",
-    "systems/dnd5e/templates/items/parts/item-spellcasting.hbs",
-    "systems/dnd5e/templates/items/parts/item-source.hbs",
     "systems/dnd5e/templates/items/parts/item-summary.hbs",
     "systems/dnd5e/templates/items/parts/item-tooltip.hbs",
     "systems/dnd5e/templates/items/parts/spell-block.hbs",
