--- templates/items/container.hbs
@@ -1,116 +1,0 @@
-<form class="{{cssClass}} flexcol" autocomplete="off">
-
-    {{!-- Item Sheet Header --}}
-    <header class="sheet-header flexrow">
-        <img class="profile" src="{{item.img}}" data-tooltip="{{item.name}}" data-edit="img">
-
-        <div class="header-details flexrow">
-            <h1 class="charname">
-                {{#if system.identified}}
-                <input name="name" type="text" value="{{item.name}}" placeholder="{{ localize 'DND5E.ItemName' }}">
-                {{else}}
-                <input name="system.unidentified.name" type="text" value="{{system.unidentified.name}}"
-                       placeholder="{{ localize 'DND5E.NameUnidentified' }}">
-                {{/if}}
-            </h1>
-
-            <div class="item-subtitle">
-                <h4 class="item-type">{{itemType}}</h4>
-                <label class="equipped">
-                    <input type="checkbox" name="system.equipped" {{checked system.equipped}}>
-                    {{#if system.equipped}}
-                        {{ localize "DND5E.Equipped" }}
-                        {{#if editable}}<i class="fa-solid fa-toggle-on"></i>{{/if}}
-                    {{else}}
-                        {{ localize "DND5E.Unequipped" }}
-                        {{#if editable}}<i class="fa-solid fa-toggle-off"></i>{{/if}}
-                    {{/if}}
-                </label>
-                <label class="identified">
-                    <input type="checkbox" name="system.identified" {{checked system.identified}}>
-                    {{#if system.identified}}
-                    {{ localize "DND5E.Identified" }}
-                    {{#if editable}}<i class="fa-solid fa-toggle-on"></i>{{/if}}
-                    {{else}}
-                    {{ localize "DND5E.Unidentified.Title" }}
-                    {{#if editable}}<i class="fa-solid fa-toggle-off"></i>{{/if}}
-                    {{/if}}
-                </label>
-            </div>
-
-            <ul class="summary flexrow">
-                <li>
-                    {{#if concealDetails}}
-                    <span>{{ localize "DND5E.Unidentified.Title" }}</span>
-                    {{else}}
-                    <select name="system.rarity" {{~#if concealDetails}} inert{{/if}}>
-                        {{selectOptions config.itemRarity selected=system.rarity blank=""}}
-                    </select>
-                    {{/if}}
-                </li>
-                <li>
-                    {{> "dnd5e.item-source"}}
-                </li>
-            </ul>
-        </div>
-    </header>
-
-    {{!-- Item Sheet Navigation --}}
-    <nav class="sheet-navigation tabs" data-group="primary">
-        <a class="item" data-tab="contents">{{localize "DND5E.Contents"}}</a>
-        <a class="item" data-tab="description">{{localize "DND5E.Description"}}</a>
-        <a class="item" data-tab="details">{{localize "DND5E.Details"}}</a>
-    </nav>
-
-    {{!-- Item Sheet Body --}}
-    <section class="sheet-body">
-        {{!-- Contents Tab --}}
-        <div class="tab inventory flexcol" data-group="primary" data-tab="contents">
-            {{> "dnd5e.inventory" sections=inventory}}
-        </div>
-
-        {{!-- Description Tab --}}
-        {{> "dnd5e.item-description"}}
-
-        {{!-- Details Tab --}}
-        <div class="tab details" data-group="primary" data-tab="details">
-            {{#dnd5e-concealSection concealDetails}}
-
-            <h3 class="form-header">{{localize "DND5E.ItemContainerDetails"}}</h3>
-
-            {{!-- Container Properties --}}
-            <div class="form-group stacked container-properties">
-                <label>{{ localize "DND5E.ItemContainerProperties" }}</label>
-                {{#each properties}}
-                <label class="checkbox">
-                    <input type="checkbox" name="system.properties.{{ @key }}" {{ checked selected }}>
-                    {{ label }}
-                </label>
-                {{/each}}
-            </div>
-
-            <div class="form-group">
-                <label>{{localize "DND5E.ItemContainerCapacity"}}</label>
-                <div class="form-fields">
-                    {{numberInput system.capacity.value name="system.capacity.value" placeholder="—"}}
-                </div>
-            </div>
-
-            <div class="form-group">
-                <label>{{localize "DND5E.ItemContainerCapacityType"}}</label>
-                <select name="system.capacity.type">
-                    {{selectOptions config.itemCapacityTypes selected=system.capacity.type}}
-                </select>
-            </div>
-
-            <div class="form-group">
-                <label>{{localize "DND5E.Attunement"}}</label>
-                <select name="system.attunement" data-dtype="Number">
-                    {{selectOptions config.attunements selected=system.attunement}}
-                </select>
-            </div>
-
-            {{/dnd5e-concealSection}}
-        </div>
-    </section>
-</form>
