--- templates/items/parts/item-mountable.hbs
@@ -1,16 +1,0 @@
-<div class="form-group">
-    <label>{{localize 'DND5E.HitPoints'}}</label>
-    <div class="form-fields">
-        {{numberInput system.hp.value name="system.hp.value" placeholder="0"}}
-        <span class="sep">&sol;</span>
-        {{numberInput system.hp.max name="system.hp.max" placeholder="0"}}
-        {{numberInput system.hp.dt name="system.hp.dt" placeholder=(localize "DND5E.Threshold")}}
-    </div>
-</div>
-
-<div class="form-group">
-    <label>{{localize 'DND5E.VEHICLE.MOUNTABLE.FIELDS.hp.conditions.label'}}</label>
-    <div class="form-fields">
-        <input type="text" name="system.hp.conditions" value="{{system.hp.conditions}}">
-    </div>
-</div>
