--- templates/items/race.hbs
@@ -1,92 +1,0 @@
-<form class="{{cssClass}} flexcol" autocomplete="off">
-    {{!-- Item Sheet Header --}}
-    <header class="sheet-header flexrow">
-        <img class="profile" src="{{item.img}}" title="{{item.name}}" data-edit="img">
-
-        <div class="header-details flexrow">
-            <h1 class="charname">
-                <input name="name" type="text" value="{{item.name}}" placeholder="{{localize 'DND5E.Species.Name'}}">
-            </h1>
-
-            <div class="item-subtitle">
-                <h4 class="item-type">{{itemType}}</h4>
-                <span class="item-status">{{itemStatus}}</span>
-            </div>
-
-            <ul class="summary flexrow">
-                <li>
-                    {{> "dnd5e.item-source"}}
-                </li>
-            </ul>
-        </div>
-    </header>
-
-    {{!-- Item Sheet Navigation --}}
-    <nav class="sheet-navigation tabs" data-group="primary">
-        <a class="item active" data-tab="description">{{localize "DND5E.Description"}}</a>
-        <a class="item" data-tab="details">{{ localize "DND5E.Details" }}</a>
-        <a class="item" data-tab="advancement">{{localize "DND5E.AdvancementTitle"}}</a>
-    </nav>
-
-    {{!-- Item Sheet Body --}}
-    <section class="sheet-body">
-
-        {{!-- Description Tab --}}
-        <div class="tab flexrow active" data-group="primary" data-tab="description">
-
-            <div class="item-properties">
-                <h4 class="properties-header">
-                    {{localize "DND5E.Type"}}
-                    {{#if editable}}<a class="config-button" data-action="type"><i class="fas fa-cog"></i></a>{{/if}}
-                </h4>
-                <ol class="properties-list">
-                    <li>{{item.system.typeLabel}}</li>
-                </ol>
-                <h4 class="properties-header">
-                    {{localize "DND5E.Movement"}} 
-                    {{#if editable}}<a class="config-button" data-action="movement"><i class="fas fa-cog"></i></a>{{/if}}
-                </h4>
-                <ol class="properties-list">
-                    {{#each item.system.movementLabels}}
-                    <li>{{this}}</li>
-                    {{else}}
-                    <li>{{localize "DND5E.None"}}</li>
-                    {{/each}}
-                </ol>
-                <h4 class="properties-header">
-                    {{localize "DND5E.Senses"}} 
-                    {{#if editable}}<a class="config-button" data-action="senses"><i class="fas fa-cog"></i></a>{{/if}}
-                </h4>
-                <ol class="properties-list">
-                    {{#each item.system.sensesLabels}}
-                    <li>{{this}}</li>
-                    {{else}}
-                    <li>{{localize "DND5E.None"}}</li>
-                    {{/each}}
-                </ol>
-            </div>
-
-            {{editor enriched.description target="system.description.value" button=true editable=editable
-                     engine="prosemirror" collaborate=false}}
-        </div>
-
-        {{!-- Details Tab --}}
-        <div class="tab details" data-group="primary" data-tab="details">
-
-            {{!-- Identifier --}}
-            <div class="form-group">
-                <label>{{ localize "DND5E.Identifier" }}</label>
-                <div class="form-fields">
-                    <input type="text" name="system.identifier" value="{{ system.identifier }}"
-                           placeholder="{{ item.identifier }}">
-                </div>
-                <p class="hint">{{ localize "DND5E.IdentifierError" }}</p>
-            </div>
-
-        </div>
-
-        {{!-- Advancement Tab --}}
-        {{> "dnd5e.item-advancement"}}
-    
-    </section>
-</form>
